Jogo de JoKenPo

+++ app/src/main/java/com/example/adrianofpinheiro/jokenpokemon/FimDeJogoActivity.kt
@@ -34,7 +34,7 @@ class FimDeJogoActivity : AppCompatActivity() {
         btJogarNovamente.setOnClickListener {
             val pontuacao = Jogador(null,
                     etNome.text.toString(),
-                    tvPontuacao.text.toString().toInt())
+                    tvPontuacao.text.toString())
 
             getJokenPokemonAPI()
                     .enviarPontos(pontuacao)
